Update example countries.json data and include Albania
There have been some changes in the data in countries.json since this
fixture was created, e.g. using full git object names, not the
abbreviated versions, and adding a 'type' property to legislatures.

This commit updates the sample countries.json data and also includes an
extra country (Albania) so we have an example of a unicameral
legislature.  The changes to the sample data necessitate some minor
changes to the URLs that are mocked, and those that are returned in the
tests.