Add a Country.houses method to find lower / upper houses
Now that legislatures in countries.json have a type property (which
seems to be either 'upper house', 'lower house' or 'unicameral
legislature', it's useful to be able to find a house by that type.

This commit adds a 'house(type_of_house)' method to Country, so you can
get the houses of that type, e.g. with:

  alabania.houses('lower house')
  >>> [<Legislature: Kuvendi in Albania>]