Add the British Virgin Islands to the test data
We're going to add some methods to return the default upper house or
lower house, which will return the house of that type with most recent
legislative period to start. The British Virgin Islands is a useful
case, since within its unicameral system, the single chamber changed its
name in 2007 from the Legislative countil to the House of Assembly,
and these are represented by different legislatures in the
EveryPolitician data.